Rename actionTemplates
