+++ _common/gameModel/libs/ClientScript/UIfacade/actionFacade.ts
@@ -37,11 +37,11 @@ import {
   CasuMessageTemplate,
   PretriageReportTemplate,
   SendRadioMessageTemplate,
-} from '../game/common/actions/actionTemplate/actionTemplateBaseRadio';
+} from '../game/common/actions/actionTemplate/actionTemplateRadio';
 import {
   MoveActorActionTemplate,
   SituationUpdateActionTemplate,
-} from '../game/common/actions/actionTemplate/actionTemplateBaseActors';
+} from '../game/common/actions/actionTemplate/actionTemplateActors';
 
 // used in page 45 (actionStandardList)
 export function getAvailableActionTemplates(
